Fix return data type of ascent
The ascent function returns a data type int64. However, the minimum and maximum values of the image array are 0 and 255, respectively. The data type should be uint8.

+++ scipy/misc/_common.py
@@ -178,10 +178,11 @@ def ascent():
     """
     import pickle
     import os
+    import numpy as np
     fname = os.path.join(os.path.dirname(__file__),'ascent.dat')
     with open(fname, 'rb') as f:
         ascent = array(pickle.load(f))
-    return ascent
+    return ascent.astype(np.uint8)
 
 
 def face(gray=False):
